+++ packages/frontend/src/pages/Settings.tsx
@@ -1,5 +1,5 @@
-import { Collapse } from '@blueprintjs/core';
-import React, { FC, useMemo, useState } from 'react';
+import { Collapse, NonIdealState } from '@blueprintjs/core';
+import { FC, useState } from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 import Content from '../components/common/Page/Content';
 import PageWithSidebar from '../components/common/Page/PageWithSidebar';
@@ -16,7 +16,10 @@ import SocialLoginsPanel from '../components/UserSettings/SocialLoginsPanel';
 import UserManagementPanel from '../components/UserSettings/UserManagementPanel';
 import { useOrganisation } from '../hooks/organisation/useOrganisation';
 import { useApp } from '../providers/AppProvider';
+import { TrackPage } from '../providers/TrackingProvider';
+import { PageName } from '../types/Events';
 import { PasswordRecoveryForm } from './PasswordRecoveryForm';
+import ProjectSettings from './ProjectSettings';
 import {
     CardContainer,
     CollapseTrigger,
@@ -50,20 +53,49 @@ const ExpandableSection: FC<ExpandableSectionProps> = ({ label, children }) => {
 };
 
 const Settings: FC = () => {
-    const { health, user } = useApp();
-    const allowPasswordAuthentication =
-        !health.data?.auth.disablePasswordAuthentication;
-    const { data: orgData, isLoading } = useOrganisation();
+    const {
+        health: {
+            data: health,
+            isLoading: isHealthLoading,
+            error: healthError,
+        },
+        user: { data: user, isLoading: isUserLoading, error: userError },
+    } = useApp();
+    const {
+        data: organization,
+        isLoading: isOrganizationLoading,
+        error: organizationError,
+    } = useOrganisation();
 
-    const basePath = useMemo(() => `/generalSettings`, []);
-
-    if (isLoading) {
+    if (isHealthLoading || isUserLoading || isOrganizationLoading) {
         return <PageSpinner />;
     }
 
+    if (userError || healthError || organizationError) {
+        return (
+            <div style={{ marginTop: '20px' }}>
+                <NonIdealState
+                    title="Unexpected error"
+                    description={
+                        userError?.error.message ||
+                        healthError?.error.message ||
+                        organizationError?.error.message
+                    }
+                />
+            </div>
+        );
+    }
+
+    if (!health || !user || !organization) return null;
+
+    const basePath = `/generalSettings`;
+
+    const allowPasswordAuthentication =
+        !health.auth.disablePasswordAuthentication;
+
     const hasSocialLogin =
-        health.data?.auth.google.oauth2ClientId ||
-        health.data?.auth.okta.enabled;
+        health.auth.google.oauth2ClientId || health.auth.okta.enabled;
+
     return (
         <PageWithSidebar>
             <Sidebar title="Settings" noMargin>
@@ -77,8 +109,8 @@ const Settings: FC = () => {
                                 to={`${basePath}/password`}
                             />
                         )}
-                        {(health.data?.auth.google.oauth2ClientId ||
-                            health.data?.auth.okta.enabled) && (
+                        {(health.auth.google.oauth2ClientId ||
+                            health.auth.okta.enabled) && (
                             <RouterMenuItem
                                 text="Social logins"
                                 exact
@@ -92,14 +124,14 @@ const Settings: FC = () => {
                         />
                     </ExpandableSection>
                     <ExpandableSection label="Organization settings">
-                        {user.data?.ability?.can('manage', 'Organization') && (
+                        {user.ability.can('manage', 'Organization') && (
                             <RouterMenuItem
                                 text="Organization"
                                 exact
                                 to={`${basePath}/organization`}
                             />
                         )}
-                        {user.data?.ability?.can(
+                        {user.ability.can(
                             'view',
                             'OrganizationMemberProfile',
                         ) && (
@@ -108,15 +140,16 @@ const Settings: FC = () => {
                                 to={`${basePath}/userManagement`}
                             />
                         )}
-                        {orgData &&
-                            !orgData.needsProject &&
-                            user.data?.ability?.can('view', 'Project') && (
+
+                        {organization &&
+                            !organization.needsProject &&
+                            user.ability.can('view', 'Project') && (
                                 <RouterMenuItem
                                     text="Project management"
-                                    exact
                                     to={`${basePath}/projectManagement`}
                                 />
                             )}
+
                         <RouterMenuItem
                             text="Appearance"
                             exact
@@ -153,7 +186,8 @@ const Settings: FC = () => {
                         </Content>
                     </Route>
                 )}
-                {user.data?.ability?.can('manage', 'Organization') && (
+
+                {user.ability.can('manage', 'Organization') && (
                     <Route exact path={`/generalSettings/organization`}>
                         <Content>
                             <CardContainer>
@@ -163,10 +197,8 @@ const Settings: FC = () => {
                         </Content>
                     </Route>
                 )}
-                {user.data?.ability?.can(
-                    'view',
-                    'OrganizationMemberProfile',
-                ) && (
+
+                {user.ability.can('view', 'OrganizationMemberProfile') && (
                     <Route path={`/generalSettings/userManagement`}>
                         <Content>
                             <ContentWrapper>
@@ -175,9 +207,10 @@ const Settings: FC = () => {
                         </Content>
                     </Route>
                 )}
-                {orgData &&
-                    !orgData.needsProject &&
-                    user.data?.ability?.can('view', 'Project') && (
+
+                {organization &&
+                    !organization.needsProject &&
+                    user.ability.can('view', 'Project') && (
                         <Route
                             exact
                             path={`/generalSettings/projectManagement`}
@@ -190,6 +223,26 @@ const Settings: FC = () => {
                         </Route>
                     )}
 
+                {organization &&
+                    !organization.needsProject &&
+                    user.ability.can('view', 'Project') && (
+                        <Route
+                            exact
+                            path={[
+                                '/generalSettings/projectManagement/:projectUuid/:tab?',
+                                '/generalSettings/projectManagement/:projectUuid/integrations/:tab',
+                            ]}
+                        >
+                            <TrackPage name={PageName.PROJECT_SETTINGS}>
+                                <Content>
+                                    <ContentWrapper>
+                                        <ProjectSettings />
+                                    </ContentWrapper>
+                                </Content>
+                            </TrackPage>
+                        </Route>
+                    )}
+
                 <Route exact path={`/generalSettings/appearance`}>
                     <Content>
                         <CardContainer>
@@ -198,6 +251,7 @@ const Settings: FC = () => {
                         </CardContainer>
                     </Content>
                 </Route>
+
                 <Route exact path={`/generalSettings/personalAccessTokens`}>
                     <Content>
                         <ContentWrapper>
@@ -205,6 +259,7 @@ const Settings: FC = () => {
                         </ContentWrapper>
                     </Content>
                 </Route>
+
                 <Route exact path={`/generalSettings`}>
                     <Content>
                         <CardContainer>
@@ -213,6 +268,7 @@ const Settings: FC = () => {
                         </CardContainer>
                     </Content>
                 </Route>
+
                 <Redirect to={basePath} />
             </Switch>
         </PageWithSidebar>